Make each minion task to a separated package
