--- pinot-plugins/pinot-minion-tasks/pinot-minion-builtin-tasks/src/main/java/org/apache/pinot/plugin/minion/executor/ConvertToRawIndexTaskExecutor.java
@@ -1,52 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.plugin.minion.executor;
-
-import java.io.File;
-import java.util.Collections;
-import java.util.Map;
-import org.apache.pinot.common.metadata.segment.SegmentZKMetadataCustomMapModifier;
-import org.apache.pinot.core.common.MinionConstants;
-import org.apache.pinot.core.minion.PinotTaskConfig;
-import org.apache.pinot.core.minion.RawIndexConverter;
-import org.apache.pinot.spi.utils.builder.TableNameBuilder;
-
-
-public class ConvertToRawIndexTaskExecutor extends BaseSingleSegmentConversionExecutor {
-
-  @Override
-  protected SegmentConversionResult convert(PinotTaskConfig pinotTaskConfig, File indexDir, File workingDir)
-      throws Exception {
-    Map<String, String> configs = pinotTaskConfig.getConfigs();
-    String tableNameWithType = configs.get(MinionConstants.TABLE_NAME_KEY);
-    String rawTableName = TableNameBuilder.extractRawTableName(tableNameWithType);
-    new RawIndexConverter(rawTableName, indexDir, workingDir,
-        configs.get(MinionConstants.ConvertToRawIndexTask.COLUMNS_TO_CONVERT_KEY)).convert();
-    return new SegmentConversionResult.Builder().setFile(workingDir)
-        .setTableNameWithType(configs.get(MinionConstants.TABLE_NAME_KEY))
-        .setSegmentName(configs.get(MinionConstants.SEGMENT_NAME_KEY)).build();
-  }
-
-  @Override
-  protected SegmentZKMetadataCustomMapModifier getSegmentZKMetadataCustomMapModifier() {
-    return new SegmentZKMetadataCustomMapModifier(SegmentZKMetadataCustomMapModifier.ModifyMode.UPDATE, Collections
-        .singletonMap(MinionConstants.ConvertToRawIndexTask.TASK_TYPE + MinionConstants.TASK_TIME_SUFFIX,
-            String.valueOf(System.currentTimeMillis())));
-  }
-}
